03b_Agent_Evaluation_Done/backend/requirements.txt
google-adk[eval]>=1.3.0
fastapi==0.115.0
uvicorn[standard]>=0.34.0
watchfiles==1.1.0
Pillow==10.4.0
python-dotenv==1.0.1
websockets==13.1
gunicorn==22.0.0
python-multipart==0.0.12
certifi==2024.8.30
google-cloud-aiplatform[adk,agent_engines]>=1.38.0
google-cloud-storage>=2.16.0
google-genai>=0.2.0 